Update participants.php
The while loop in the file was changed. Instead of searching for the student number of the current user in the enrolment table, we are searching for the COURSE CODE that the user is currently viewing. Then, we get the row(s) of the students enrolled in that course and search for their USER ID in the registration table to get their first and last name. I added a new column to also display their profile pic in the table.