Update course.php
A link was added to the "Participants" text which directs the user to the participants page. In the link, we add an ID (id= course code) that way we can use the $_GET request from the participants page to get the course code of the current course that is being viewed.