Add files via upload
The announcement.php file has been updated. The user can now create an announcement.
The create_course.php and the edit_course.php has also been updated to accommodate the new changes on database for the courses table.
The changes on the course.php  are just the links to the participants.php and announcement.php in the navigation bar.